Fixing bug where ID was being set to empty string
IDs were set to empty strings which was causing javascript errors since
templated code resolved into the form:

var something = ;

Where the templated variable was printed before the semicolon, but since
it was an empty string javascript would halt with an error.  The empty
strings have been replaced with a 0 instead.